requirements.txt
black
botorch
cheetah-accelerator>=0.8.0
flake8
isort
jupyter
lightning
matplotlib
numba
numexpr
numpy
git+https://github.com/ocelot-collab/ocelot.git
pyyaml
retry
SciencePlots
scikit-learn
seaborn
torch
tqdm
xopt>=2.6.7
